+++ packages/backend/src/models/Meta.ts
@@ -570,6 +570,23 @@ export class MiMeta {
 	public bannedEmailDomains: string[];
 
 	@Column('varchar', {
+		length: 1024,
+		array: true,
+		default: '{}',
+	})
+	public allowedEmailDomains: string[];
+
+	@Column('boolean', {
+		default: false,
+	})
+	public enableAutoAddBannedEmailDomain: boolean;
+
+	@Column('boolean', {
+		default: false,
+	})
+	public enableAllowedEmailDomainsOnly: boolean;
+
+	@Column('varchar', {
 		length: 1024, array: true, default: '{ "admin", "administrator", "root", "system", "maintainer", "host", "mod", "moderator", "owner", "superuser", "staff", "auth", "i", "me", "everyone", "all", "mention", "mentions", "example", "user", "users", "account", "accounts", "official", "help", "helps", "support", "supports", "info", "information", "informations", "announce", "announces", "announcement", "announcements", "notice", "notification", "notifications", "dev", "developer", "developers", "tech", "misskey" }',
 	})
 	public preservedUsernames: string[];