+++ packages/backend/src/server/api/endpoints/admin/update-meta.ts
@@ -141,6 +141,9 @@ export const paramDef = {
 		enableIdenticonGeneration: { type: 'boolean' },
 		serverRules: { type: 'array', items: { type: 'string' } },
 		bannedEmailDomains: { type: 'array', items: { type: 'string' } },
+		allowedEmailDomains: { type: 'array', items: { type: 'string' } },
+		enableAutoAddBannedEmailDomain: { type: 'boolean' },
+		enableAllowedEmailDomainsOnly: { type: 'boolean' },
 		preservedUsernames: { type: 'array', items: { type: 'string' } },
 		manifestJsonOverride: { type: 'string' },
 		enableFanoutTimeline: { type: 'boolean' },
@@ -637,6 +640,18 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 
 			if (ps.bannedEmailDomains !== undefined) {
 				set.bannedEmailDomains = ps.bannedEmailDomains;
+			}
+
+			if (ps.allowedEmailDomains !== undefined) {
+				set.allowedEmailDomains = ps.allowedEmailDomains;
+			}
+
+			if (ps.enableAutoAddBannedEmailDomain !== undefined) {
+				set.enableAutoAddBannedEmailDomain = ps.enableAutoAddBannedEmailDomain;
+			}
+
+			if (ps.enableAllowedEmailDomainsOnly !== undefined) {
+				set.enableAllowedEmailDomainsOnly = ps.enableAllowedEmailDomainsOnly;
 			}
 
 			if (ps.urlPreviewEnabled !== undefined) {